feat: cc-0120-finished my order ssr page functionality
