+++ src/components/NavBar/Profile.tsx
@@ -2,12 +2,11 @@ import { Menu, MenuButton, MenuList, MenuItem, Flex } from "@chakra-ui/react";
 import { FaRegUser } from "react-icons/fa";
 import PROFILE_LISTS from "@/constants/profileLists";
 import { BiChevronDown } from "react-icons/bi";
-import {
-  switchMyTemplateDisplay,
-  switchMyOrderDisplay,
-  switchMyAccount,
-} from "@/store/reducer/buttonToggleSlice";
+import { switchMyTemplateDisplay, switchMyAccount } from "@/store/reducer/buttonToggleSlice";
+import { userData } from "@/store/reducer/userSlice";
+import { useStoreSelector } from "@/store/hooks";
 import { useDispatch } from "react-redux";
+import { useRouter } from "next/router";
 
 interface Props {
   isOpen: boolean;
@@ -17,24 +16,32 @@ interface Props {
 }
 
 const Profile = ({ isOpen, onOpen, onClose, handleLogout }: Props) => {
+  const currentUserId = useStoreSelector(userData).userId;
   const disPatch = useDispatch();
-
+  const router = useRouter();
   const onClickHandler = (title: string) => {
     switch (title) {
       case "My Account":
+        router.push("/");
         disPatch(switchMyAccount(true));
         onClose();
         return;
       case "My Order":
-        disPatch(switchMyOrderDisplay(true));
+        router.push(`/my_order?user_id=${currentUserId}`);
         onClose();
         return;
       case "My Template":
+        router.push("/");
         disPatch(switchMyTemplateDisplay(true));
         onClose();
         return;
       case "Sign Out":
+        router.push("/");
+        disPatch(switchMyAccount(false));
+        disPatch(switchMyTemplateDisplay(false));
         handleLogout();
+        onClose();
+
         return;
       default:
         return;
