+++ src/constants/index.ts
@@ -3,4 +3,5 @@ export const PAGE_NOT_FOUND = "/404";
 export const PAYMENT = "/payment";
 export const TEAM_PAGE = "/team";
 export const TEAM_PAGE_PERFORMANCE_TESTING = "/team1";
+export const MY_ORDER = "/my_order";
 export default HOME_PAGE_LINK;
