+++ src/layouts/index.tsx
@@ -9,12 +9,12 @@ import PriceBar from "@/components/PriceBar";
 import TopBar from "@/components/TopBar";
 import ShoppingCart from "@/components/ShoppingCart";
 import MyTemplate from "@/components/MyTemplate";
-import MyOrder from "@/components/MyOrder";
 import {
   PAGE_NOT_FOUND,
   TEAM_PAGE,
   TEAM_PAGE_PERFORMANCE_TESTING,
   PAYMENT,
+  MY_ORDER,
 } from "../../src/constants";
 
 import OrderGeneration from "@/components/OrderGeneration";
@@ -50,6 +50,16 @@ const Layout: React.FC<{ children: ReactNode }> = ({ children }) => {
           </Box>
         </>
       );
+    case MY_ORDER:
+      return (
+        <>
+          <Header />
+          <Box as="main">
+            <NavigationBar />
+            {children}
+          </Box>
+        </>
+      );
 
     case PAYMENT:
       return (
@@ -68,7 +78,6 @@ const Layout: React.FC<{ children: ReactNode }> = ({ children }) => {
           <Box as="main">
             <NavigationBar />
             <MyTemplate />
-            <MyOrder />
             <MyAccount />
             <ShoppingCart />
             <OrderGeneration />
