+++ src/store/reducer/buttonToggleSlice.ts
@@ -11,7 +11,6 @@ export interface ButtonToggleState {
   isTemplateSelect: boolean;
   isOrderGenerationOpen: boolean;
   isMyTemplateOpen: boolean;
-  isMyOrderOpen: boolean;
   isSwitch3D: boolean;
   isMyAccountOpen: boolean;
 }
@@ -27,7 +26,6 @@ export const initialState: ButtonToggleState = {
   isTemplateSelect: false,
   isOrderGenerationOpen: false,
   isMyTemplateOpen: false,
-  isMyOrderOpen: false,
   isSwitch3D: false,
   isMyAccountOpen: false,
 };
@@ -49,7 +47,6 @@ export const ButtonToggleSlice = createSlice({
         isSideBarOpen: false,
         isOrderGenerationOpen: false,
         isMyTemplateOpen: false,
-        isMyOrderOpen: false,
         isSwitch3D: false,
         isMyAccountOpen: false,
       };
@@ -91,7 +88,6 @@ export const ButtonToggleSlice = createSlice({
         isSideBarOpen: false,
         isOrderGenerationOpen: false,
         isMyTemplateOpen: false,
-        isMyOrderOpen: false,
         isSwitch3D: false,
         isMyAccountOpen: false,
       };
@@ -106,21 +102,10 @@ export const ButtonToggleSlice = createSlice({
       return {
         ...state,
         isMyTemplateOpen: action.payload,
-        isMyOrderOpen: false,
-        isCartOpen: false,
-        isOrderGenerationOpen: false,
-        isSwitch3D: false,
-      };
-    },
-    switchMyOrderDisplay: (state: ButtonToggleState, action: PayloadAction<boolean>) => {
-      return {
-        ...state,
-        isMyOrderOpen: action.payload,
-        isMyTemplateOpen: false,
-        isCartOpen: false,
-        isOrderGenerationOpen: false,
-        isSwitch3D: false,
         isMyAccountOpen: false,
+        isCartOpen: false,
+        isOrderGenerationOpen: false,
+        isSwitch3D: false,
       };
     },
     startSelectTemplate: (state: ButtonToggleState, action: PayloadAction<boolean>) => {
@@ -165,7 +150,6 @@ export const {
   switchLoginModal,
   switchCreateTemplate,
   switchMyTemplateDisplay,
-  switchMyOrderDisplay,
   startSelectTemplate,
   switchOrderGeneration,
   switch3D,
